html/semantics/embedded-content/the-iframe-element/sandbox_006_bidi.htm

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>Test console log are present</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<iframe src="support/iframe_sandbox_006.htm" sandbox="allow-forms"></iframe>
<script type="text/javascript">
    promise_test(async () => {
        const tree = await test_driver.bidi.browsing_context.get_tree({root: window});
        const iframe = tree[0].children[0];
        const submit_button_selector_result = await test_driver.bidi.browsing_context.locate_nodes({
            context: iframe.context,
            locator: {type: "css", value: "#submitButton"}
        });
        const submit_button_shared_id = submit_button_selector_result.nodes[0].sharedId;
        await test_driver.bidi.input.click({
            context: iframe.context,
            origin: {
                type: "element",
                element: {
                    sharedId: submit_button_shared_id
                }
            }
        });

        // Wait is required to avoid false-negative.
        // TODO: replace wait with a more efficient method.
        await new Promise(resolve => setTimeout(resolve, 300));

        const success_message_selector_result = await test_driver.bidi.browsing_context.locate_nodes({
            context: iframe.context,
            locator: {type: "innerText", value: "PASS!!!"}
        });

        assert_equals(success_message_selector_result.nodes.length, 1);
        assert_equals(success_message_selector_result.nodes[0].value.attributes.style, 'color: Green');
    }, "Allow form submission inside sandbox iframe when sandbox='allow-forms'.");
</script>
